refactor(reviewer-bot): finish events cleanup
